Addition: add new feature which allows player to swap out learned moves per individual Pokémon. Also add basic UI/UX to allow user to swap/rearrange moves.
Description: All moves learned by the Pokémon will accumulate and are shown in a list. Moves that are active are contained in a different list than moves that are learned, but inactive. Players can swap and rearrange moves as they please, swapping active moves for inactive ones and vice versa.
Edits made to: party menu.